--- frontend/components/ui/badge/index.ts
@@ -1,2 +1,0 @@
-export { Badge } from "./badge";
-export { badgeVariants } from "./variants";
